+++ src/theme.js
@@ -1,166 +1,185 @@
-import { light } from '@eva-design/eva';
+import { DefaultTheme } from '@react-navigation/native';
 
-export const theme = {
-  ...light,
-  'color-primary-default': '#1F93FF',
-  'color-white': '#FFFFFF',
+export const palette = {
+  background: '#FFFFFF',
+  backgroundLight: '#F8FAFC',
+  backgroundDark: '#E4EBF1',
 
-  'color-basic-transparent-300': '$color-primary-500',
-  'color-primary-focus': '$color-primary-default',
-  'color-primary-active': '$color-primary-default',
+  backdropColor: 'rgba(0, 0, 0, 0.5)',
 
-  'color-primary-50': '#EBF5FF',
-  'color-primary-75': '#D6EBFF',
-  'color-primary-100': '#C2E1FF',
-  'color-primary-200': '#99CEFF',
-  'color-primary-300': '#70BAFF',
-  'color-primary-400': '#47A6FF',
-  'color-primary-500': '#1F93FF',
-  'color-primary-600': '#1976CC',
-  'color-primary-700': '#135899',
-  'color-primary-800': '#0C3B66',
-  'color-primary-900': '#061D33',
+  backgroundChat: '#f4f6fb',
+  backgroundPrivate: '#FFEDBF',
+  backgroundPrivateLight: '#FFF4D9',
+  backgroundDate: '#e1f5feeb',
+  backgroundActivity: '#e7eefb',
 
-  'color-primary-transparent-100': 'rgba(31, 147, 255, 0.08)',
-  'color-primary-transparent-200': 'rgba(31, 147, 255, 0.16)',
-  'color-primary-transparent-300': 'rgba(31, 147, 255, 0.24)',
-  'color-primary-transparent-400': 'rgba(31, 147, 255, 0.32)',
-  'color-primary-transparent-500': 'rgba(31, 147, 255, 0.4)',
-  'color-primary-transparent-600': 'rgba(31, 147, 255, 0.48)',
+  primaryColor: '#1F93FF',
+  primaryColorDark: '#1976CC',
+  primaryColorDarker: '#135899',
+  primaryColorLight: '#EBF5FF',
 
-  'color-secondary-50': '#F1F5F8',
-  'color-secondary-75': '#EBF0F5',
-  'color-secondary-100': '#E4EBF1',
-  'color-secondary-200': '#C9D7E3',
-  'color-secondary-300': '#AEC3D5',
-  'color-secondary-400': '#93AFC8',
-  'color-secondary-500': '#779BBB',
-  'color-secondary-600': '#446888',
-  'color-secondary-700': '#37546D',
-  'color-secondary-800': '#293F51',
-  'color-secondary-900': '#1B2836',
+  infoColor: '#077BFF',
+  infoColorDark: '#055FDB',
+  infoColorDarker: '#0346B7',
+  infoColorLight: '#9BD8FF',
+  infoColorLighter: '#C2E1FF',
 
-  'color-black-100': '#F2F3F7',
-  'color-black-200': '#E9EAEF',
-  'color-black-300': '#DADBDF',
-  'color-black-400': '#B6B7BB',
-  'color-black-500': '#96979C',
-  'color-black-600': '#6E6F73',
-  'color-black-700': '#5A5B5F',
-  'color-black-800': '#3C3D40',
-  'color-black-900': '#1B1C1F',
+  secondaryColor: '#5d7592',
+  secondaryColorLight: '#C9D7E3',
+  secondaryColorDarker: '#293F51',
 
-  'color-success-100': '#CEFCD0',
-  'color-success-200': '#9FFAAC',
-  'color-success-300': '#6DF08D',
-  'color-success-400': '#48E17C',
-  'color-success-500': '#13ce66',
-  'color-success-600': '#0DB166',
-  'color-success-700': '#099462',
-  'color-success-800': '#06775A',
-  'color-success-900': '#036253',
+  successColor: '#44CE4B',
+  violetColor: '#AC52FF',
 
-  'color-success-transparent-100': 'rgba(19, 206, 102, 0.08)',
-  'color-success-transparent-200': 'rgba(19, 206, 102, 0.16)',
-  'color-success-transparent-300': 'rgba(19, 206, 102, 0.24)',
-  'color-success-transparent-400': 'rgba(19, 206, 102, 0.32)',
-  'color-success-transparent-500': 'rgba(19, 206, 102, 0.4)',
-  'color-success-transparent-600': 'rgba(19, 206, 102, 0.48)',
+  dangerColor: '#ff382d',
+  dangerColorDark: '#ed2f2f',
+  dangerColorDarker: '#930F1F',
+  dangerColorLight: '#FFCCD1',
 
-  'color-info-75': '#C2E1FF',
-  'color-info-100': '#CDEDFF',
-  'color-info-200': '#9BD8FF',
-  'color-info-300': '#6ABDFF',
-  'color-info-400': '#45A4FF',
-  'color-info-500': '#077BFF',
-  'color-info-600': '#055FDB',
-  'color-info-700': '#0346B7',
-  'color-info-800': '#023193',
-  'color-info-900': '#01237A',
+  warningColor: '#ffc532',
+  warningColorLight: '#FFEDBF',
+  warningColorDarker: '#7A4D09',
 
-  'color-info-transparent-100': 'rgba(7, 123, 255, 0.08)',
-  'color-info-transparent-200': 'rgba(7, 123, 255, 0.16)',
-  'color-info-transparent-300': 'rgba(7, 123, 255, 0.24)',
-  'color-info-transparent-400': 'rgba(7, 123, 255, 0.32)',
-  'color-info-transparent-500': 'rgba(7, 123, 255, 0.4)',
-  'color-info-transparent-600': 'rgba(7, 123, 255, 0.48)',
+  colorWhite: '#FFFFFF',
 
-  'color-warning-100': '#FFF8D6',
-  'color-warning-200': '#FFEEAD',
-  'color-warning-300': '#FFE383',
-  'color-warning-400': '#FFD765',
-  'color-warning-500': '#FFC532',
-  'color-warning-600': '#DBA224',
-  'color-warning-700': '#B78119',
-  'color-warning-800': '#93630F',
-  'color-warning-900': '#7A4D09',
+  colorBlack: '#000000',
+  colorBlackLight: '#96979C',
 
-  'color-warning-transparent-100': 'rgba(255, 197, 50, 0.08)',
-  'color-warning-transparent-200': 'rgba(255, 197, 50, 0.16)',
-  'color-warning-transparent-300': 'rgba(255, 197, 50, 0.24)',
-  'color-warning-transparent-400': 'rgba(255, 197, 50, 0.32)',
-  'color-warning-transparent-500': 'rgba(255, 197, 50, 0.4)',
-  'color-warning-transparent-600': 'rgba(255, 197, 50, 0.48)',
+  textLighter: '#779bbb',
+  textLight: '#446888',
+  text: '#37546D',
+  textDark: '#293F51',
+  textDarker: '#1f2d3d',
 
-  'color-danger-100': '#FFE7D5',
-  'color-danger-200': '#FFC9AC',
-  'color-danger-300': '#FFA582',
-  'color-danger-400': '#FF8263',
-  'color-danger-500': '#FF4830',
-  'color-danger-600': '#DB2923',
-  'color-danger-700': '#B7181F',
-  'color-danger-800': '#930F1F',
-  'color-danger-900': '#ed2f2f',
+  borderDark: '#37546D',
+  border: '#C9D7E3',
+  borderLight: '#EBF0F5',
 
-  'color-danger-transparent-100': 'rgba(255, 72, 48, 0.08)',
-  'color-danger-transparent-200': 'rgba(255, 72, 48, 0.16)',
-  'color-danger-transparent-300': 'rgba(255, 72, 48, 0.24)',
-  'color-danger-transparent-400': 'rgba(255, 72, 48, 0.32)',
-  'color-danger-transparent-500': 'rgba(255, 72, 48, 0.4)',
-  'color-danger-transparent-600': 'rgba(255, 72, 48, 0.48)',
+  iconDark: '#293F51',
+  icon: '#37546D',
+  iconLight: '#446888',
+};
 
-  'color-light-gray': '$color-basic-active',
-  'color-border': '$color-basic-active',
-  'loader-color': '$color-basic-100',
+export const spacing = {
+  zero: 0,
+  tiny: 2,
+  micro: 4,
+  smaller: 8,
+  half: 12,
+  small: 16,
+  medium: 24,
+  large: 32,
+  larger: 48,
+};
 
-  'color-background': '#f4f6fb',
-  'color-background-light': '#F9FAFC',
-  'color-background-message': '#c7e3ff',
-  'color-background-activity': '#e7eefb',
-  'color-border-activity': '#c8d6e6',
-  'color-border-private': '#ffdb7f',
-  'color-background-private': '#FFEDBF',
-  'color-background-private-light': '#FFF4D9',
+export const fontSize = {
+  xxs: 10,
+  xs: 12,
+  sm: 14,
+  md: 16,
+  lg: 18,
+  xl: 20,
+  xxl: 24,
+  xxxl: 32,
+};
+export const borderRadius = {
+  zero: 0,
+  micro: 4,
+  small: 8,
+  medium: 12,
+  large: 16,
+  larger: 24,
+  largest: 32,
+  full: 48,
+};
 
-  'color-message-left': '#3c4858',
-  'item-border-color': '#f0f4f5',
-  'color-gray': '#6e6f73',
-  'color-background-inbox': '#ebf5ff',
+export const fontWeight = {
+  thin: '100',
+  ultraLight: '200',
+  light: '300',
+  regular: '400',
+  medium: '500',
+  semiBold: '600',
+  heavy: '800',
+  bold: 'bold',
+  normal: 'normal',
+};
 
-  'color-background-date': '#e1f5feeb',
-  'text-basic-color': '#1f2d3d',
-  'text-light-color': '#3c4858',
-  'text-hint-color': '#8492a6',
-  'text-tooltip-color': '#ffffff',
+export const LightTheme = {
+  dark: false,
+  colors: {
+    ...DefaultTheme.colors,
+    ...palette,
+  },
+  spacing,
+  fontSize,
+  borderRadius,
+  fontWeight,
+};
 
-  'back-drop-color': 'rgba(0, 0, 0, 0.5)',
-  // Text color
-  'text-primary-size': 14,
-  'input-font-size': 16,
-  // Font Weight
-  'font-thin': '100',
-  'font-ultra-light': '200',
-  'font-light': '300',
-  'font-regular': '400',
-  'font-medium': '500',
-  'font-semi-bold': '600',
-  'font-bold': '700',
+export const DarkTheme = {
+  ...DefaultTheme,
+  colors: {
+    ...LightTheme.colors,
+    background: '#0D1721',
+    backgroundLight: '#1F2D3D',
+    backgroundDark: '#314155',
 
-  // Font size
-  'font-size-extra-large': 20,
-  'font-size-large': 18,
-  'font-size-medium': 16,
-  'font-size-small': 14,
-  'font-size-extra-small': 12,
-  'font-size-extra-extra-small': 10,
+    backdropColor: 'rgba(0, 0, 0, 0.5)',
+
+    backgroundChat: '#f4f6fb',
+    backgroundPrivate: '#FFEDBF',
+    backgroundPrivateLight: '#FFF4D9',
+    backgroundDate: '#e1f5feeb',
+    backgroundActivity: '#e7eefb',
+
+    primaryColor: '#1F93FF',
+    primaryColorDark: '#47A6FF',
+    primaryColorLight: '#1976CC',
+
+    infoColor: '#077BFF',
+    infoColorDark: '#055FDB',
+    infoColorDarker: '#0346B7',
+    infoColorLight: '#9BD8FF',
+    infoColorLighter: '#C2E1FF',
+
+    secondaryColor: '#5d7592',
+    secondaryColorLight: '#C9D7E3',
+    secondaryColorDarker: '#293F51',
+
+    successColor: '#00C41D',
+    violetColor: '#AC52FF',
+
+    dangerColor: '#ff382d',
+    dangerColorDark: '#ed2f2f',
+    dangerColorDarker: '#930F1F',
+    dangerColorLight: '#FFCCD1',
+
+    warningColor: '#ffc532',
+    warningColorLight: '#FFEDBF',
+    warningColorDarker: '#7A4D09',
+
+    colorWhite: '#FFFFFF',
+
+    colorBlack: '#000000',
+    colorBlackLight: '#96979C',
+
+    textLighter: '#779bbb',
+    textLight: '#C9D7E3',
+    text: '#F1F5F8',
+    textDark: '#F8FAFC',
+    textDarker: '#FFFFFF',
+
+    borderDark: '#EBF0F5',
+    border: '#C9D7E3',
+    borderLight: '#293F51',
+
+    iconDark: '#F8FAFC',
+    icon: '#F1F5F8',
+    iconLight: '#C9D7E3',
+  },
+  spacing,
+  fontSize,
+  borderRadius,
+  fontWeight,
 };
